pyproject.toml
[project]
name = "collegesystem"
version = "0.1.0"
description = ""
authors = [
    {name = "Yaneww11",email = "yaneyanev2807@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "django-unfold (>=0.59.0,<0.60.0)",
    "pillow (>=11.2.1,<12.0.0)",
]

[tool.poetry]
packages = [
    {include = "collegeSystem"}, # Need because prooject name is in lowrecase
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
